backend/requirements.txt
moviepy==1.0.3
Pillow==10.1.0
Flask-Cors==4.0.0
flask==3.0.0
